KnapSackProblemsbasics
Sample code on how 0/1 knapSack works
Given for list of items ( thier weights and thier values ) and given a knapsack, we need to include the elements such that we can maximize the profit.  Each element once picked is gone. You dont have it again.

We first use the Recursion method to solve the problem
Then we use the memozize method. 
Then we create the DP table  and avoid  the recursion completely